Database adjust

+++ src/database/migrations/20240517171612_ingredients.ts
@@ -3,7 +3,7 @@ import { Knex } from "knex";
 exports.up = (knex: Knex): Promise<void> => knex.schema.createTable("ingredients", table => {
     table.increments("id").primary();
 
-    table.integer("meal_id").references("id").inTable("meals").onDelete("CASCADE");
+    table.bigInteger("meal_id").references("id").inTable("meals").onDelete("CASCADE");
 
     table.string("name").notNullable();
